T67xx FC constants renamed, conditional including of i2c.h
For quick copy-n-paste of parameters when calling the reader of the
sensor, the FC was moved at end of the define.

Also, as the sensor can be acceded through at least 3 modes, and
in this library the sensor are only acceded with I2C, the including
of the i2c header now depends on a define in the source code.