Final multitarget and multi compilation modes
MultiCompilation Modes: in main_pic8.c can be found some compilation
modes: the main mode (SMACH, the state machine mode) and 3 testing
modes.

MultiTarget Modes: As the project is being developed through the PICs
18F2550 and 16F1769, specific files were changed to allow easy target
choosing through MPLAB X IDE. This evolves a code migration/portability
preprocessor conditional compilation based on the device selected.
The communication between the PIC and either the iM880B and CO2 sensor
was verified.

The only remaining thing is to check that, as all this changes were made
with the LoRaWAN server offline, it should be tested that the node
begins to send data after some time (with any of the PICs mentioned),
prior to merge this branch (multitarget) to master.