NMEA decoder simplified, Delay for LoRaWAN Rx windows considered
Minor cleanup of variables in some of the modified files

LoRaWAN mote example updated to use the GPS and also keep in mind
the 2 reception windows, by added a simple 2.5 s delay after
sending packets

PC unit test for GPS updated also

Most of the changes occured in NMEA.[c/h] files, due to the whole
conception of the FSM machine were simplified to simply have a
single reception buffer, thus reducing the memory consumption and
complexity of source code. The main behaviour changed is that there
is no need to poll if a NMEA sentence are pending of processing:
a handler were designed to do this on every incoming octet.